cicd/pulumi/pyproject.toml
[project]
name = "pulumi-ee-cicd"
version = "0.1.0"
description = "Pulumi code for Edge Endpoint CI/CD"
requires-python = ">=3.10"
dependencies = [
    "pulumi>=3",
    "pulumi-azure-native>=2.8.0",
    "azure-identity>=1.17.0",
    "azure-keyvault-secrets>=4.9.0",
    "groundlight>=0.22.6, <0.23.0",
    "fabric>=3.2.2",
]
